feat(falcon-client): Introducing new "graphqlProxy" flag to toggle "graphql" proxy
